Added Identity service configuration extension.
