+++ src/Web/Startup.cs
@@ -1,9 +1,9 @@
 using Masny.QRAnimal.Infrastructure.Identity;
+using Masny.QRAnimal.Web.Extensions;
 using Masny.QRAnimal.Web.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -11,8 +11,6 @@ using Microsoft.Extensions.Hosting;
 
 namespace Masny.QRAnimal.Web
 {
-
-
     public class Startup
     {
         public IConfiguration Configuration { get; }
@@ -24,30 +22,10 @@ namespace Masny.QRAnimal.Web
 
         public void ConfigureServices(IServiceCollection services)
         {
-            CreateIdentityIfNotCreated(services);
+            services.AddIdentityService();
 
             // Добавлен Identity контекст.
-            services.AddDbContext<IdentityContext>(options =>
-                options.UseSqlServer(Configuration.GetConnectionString("IdentityConnection")));
-
-            //ConfigureServices(services);
-        }
-
-        // TODO: Вынести в extension
-        private static void CreateIdentityIfNotCreated(IServiceCollection services)
-        {
-            var sp = services.BuildServiceProvider();
-            using (var scope = sp.CreateScope())
-            {
-                var existingUserManager = scope.ServiceProvider
-                    .GetService<UserManager<AppUser>>();
-                if (existingUserManager == null)
-                {
-                    services.AddIdentity<AppUser, IdentityRole>()
-                        .AddEntityFrameworkStores<IdentityContext>()
-                                        .AddDefaultTokenProviders();
-                }
-            }
+            services.AddDbContext<IdentityContext>(options => options.UseSqlServer(Configuration.GetConnectionString("IdentityConnection")));
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
@@ -62,7 +40,6 @@ namespace Masny.QRAnimal.Web
             app.UseRouting();
 
             app.UseAuthentication();
-            //app.UseAuthorization();
 
             app.UseEndpoints(endpoints =>
             {
